Some bug fixes in orders route of both user and store

+++ app/api/store/orders/route.ts
@@ -1,6 +1,7 @@
 import dbConnect from "@/Db/Db";
 import OrderModel from "@/models/Order";
 import StoreModel from "@/models/Store";
+import UserModel from "@/models/User";
 import { NextResponse } from "next/server";
 import { getServerSession, User } from "next-auth";
 import { authOptions } from "../../auth/[...nextauth]/options";
@@ -21,7 +22,7 @@ export async function GET(req:Request) {
         }
         const orders = await OrderModel.find({storeId:queryParam.storeId});
         if(orders){
-            return NextResponse.json({message:"All orders are fetched", success:true}, {status:200})
+            return NextResponse.json({message:"All orders are fetched", orders, success:true}, {status:200})
         }else{
             return NextResponse.json({message:"No orders found", success:true}, {status:200});
         }
@@ -34,25 +35,25 @@ export async function GET(req:Request) {
 // maybe the PUT to update the order status is being implemented by the debouncing like username debouncing
 export async function PUT(req:Request){
     await dbConnect();
-    const session = await getServerSession(authOptions);
-    const user:User = session?.user as User
-    if(!session || !session.user){
-        return NextResponse.json({message:"Not authenticated", success:false}, {status:401});
-    }
-    if(user.role !== "StoreOwner"){
-        return NextResponse.json({message:"Not authenticated, Not a StoreOwner", success:false}, {status:401});
-    }
     try{
+        const data = await req.json();
+        const userByEmail = await UserModel.findOne({email:data.session.user.email});
+        // checking the user is present in the db or not, if yes is he signed up as a store oner of not 
+        if(!userByEmail){
+            return NextResponse.json({message: "user not saved try to sign up again", success:false},{ status:404});
+        }else if(userByEmail.role !== 'StoreOwner'){
+            return NextResponse.json({message:"please SignUp as a Store Owner to continue.", success:false}, {status:404});
+        }
         const {searchParams} = new URL(req.url);
         const queryParam = {
             orderId:searchParams.get('orderId'),
             status:searchParams.get('status')
         }
-        const updatedOrder = OrderModel.findByIdAndUpdate(queryParam.orderId, {status:queryParam.status}, {new:true})
+        const updatedOrder = await OrderModel.findByIdAndUpdate(queryParam.orderId, {status:queryParam.status})
         if(!updatedOrder){
             return NextResponse.json({message:"Unable to update order status", success:false}, {status:400})
         }
-        return NextResponse.json({message:"Updated the order status", success:true}, {status:200})
+        return NextResponse.json({message:"Updated the order status", updatedOrder, success:true}, {status:200})
         
     }catch(err){
         return NextResponse.json({message:"Internal server error", success:false}, {status:500})
@@ -69,6 +70,13 @@ export async function DELETE(req:Request){
         // extracting the params from the query
         const queryParam = {
             orderId:searchParams.get('orderId'),
+        }
+        const order = await OrderModel.findById(queryParam.orderId);
+        if(!order){
+            return NextResponse.json({message:"Order is not found", success:false}, {status:400})
+        }
+        if(order.status !== 'Cancelled'){
+            return  NextResponse.json({message:"Cannot delete a Order which is not cancelled by the user", success:false}, {status:400})
         }
         // finding the order by Id and deleteing it
         const deleteOrder = await OrderModel.findByIdAndDelete(queryParam.orderId);
